requirements.txt
Django==5.2.7
djangorestframework==3.16.1
python-decouple==3.8
psycopg2-binary==2.9.11
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
Pillow==12.0.0
drf-spectacular==0.27.2
drf-spectacular[sidecar]==0.27.2
django-filter==24.3
